# openapi.yaml
openapi: 3.0.0
info:
  title: API Sistema Timbrature
  version: 1.0.0
  description: API per la gestione delle timbrature, turni, commesse e utenti
servers:
  - url: /api/v1
    description: Gateway BFF
paths:
  /auth/login:
    post:
      summary: Login persistente (crea sessione lunga su dispositivo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login riuscito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Credenziali errate
        '423':
          description: Account temporaneamente bloccato (rate limit)
  /auth/refresh:
    post:
      summary: Rinnovo automatico token (rotation + device binding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Nuova coppia token (refresh ruotato)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Refresh token non valido o riutilizzato (session hijack)
  /auth/logout:
    post:
      summary: Logout dal dispositivo corrente
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Sessione revocata
  /auth/logout-all:
    post:
      summary: Logout globale (revoca tutte le sessioni dell'utente)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Tutte le sessioni revocate
  /auth/me:
    get:
      summary: Verifica sessione e ritorna profilo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Utente autenticato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Token scaduto o non valido
        '403':
          description: Utente disattivato o tokenVersion non valida
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: string
      enum:
        - ADMIN
        - MANAGER
        - DIPENDENTE
    User:
      type: object
      required:
        - id
        - codiceDipendente
        - nome
        - email
        - ruolo
        - attivo
        - dataAssunzione
        - createdAt
      properties:
        id:
          type: string
        codiceDipendente:
          type: string
          description: Matricola o codice interno aziendale univoco
        nome:
          type: string
        cognome:
          type: string
        email:
          type: string
          format: email
        ruolo:
          $ref: '#/components/schemas/Role'
        reparto:
          type: string
          description: Reparto aziendale di appartenenza
        qualifica:
          type: string
          description: Qualifica o mansione (solo informativo)
        attivo:
          type: boolean
          description: Se false, l'utente non può più timbrare né accedere
        dataAssunzione:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          description: UserId dell'admin che ha creato l'account
        lastLoginAt:
          type: string
          format: date-time
        note:
          type: string
          description: Annotazioni interne (visibili solo ad admin/manager)
...
    LoginRequest:
      type: object
      required: [email, password, deviceId]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        deviceId:
          type: string
          description: ID univoco del dispositivo (persistente)
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Secondi di validità dell'access token
    RefreshRequest:
      type: object
      required: [refreshToken, deviceId]
      properties:
        refreshToken:
          type: string
        deviceId:
          type: string
    MeResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
        role:
          $ref: '#/components/schemas/Role'
        deviceId:
          type: string